+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Installation guide
+- Linear interpolation for a single point
+- Outside and keyframe buttons in the side panel for all interpolation shapes (they were only for boxes before)
 
 ### Changed
 -
@@ -20,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed incorrect width of shapes borders in some cases
 - Fixed annotation parser for tracks with a start frame less than the first segment frame
+- Fixed interpolation on the server near outside frames
 
 ### Security
 -